Added files for testing structure
